refactor: integrate item action button into button
